+++ lib/src/constants/version_constant.dart
@@ -1,2 +1,2 @@
-const String libraryVersion = "2.0.0";
+const String libraryVersion = "2.0.1";
 const String libraryType = "mobile";